Changes to PurPur blocks for cooler ships
I'm not sure where to add the screen shot. However much of the issues of Purpur blocks is that, they arent purple. They Look like a milky pink and its disgusting, this is a texture I handed around to a lot of people that allowed many to use purpur blocks in a good looking way.